Added a lockAcquired boolean to MessageTask indicating if the task holds the process lock. This is passed to the DispatchResult and used to determine if finishDispatch should release the lock. Without this, any time finishDispatch is called the lock would be released, even if that message wasn't the one who obtained the lock.
If dispatchRawMessage throws an interrupted exception while waiting for the task, it now explicitly cancels the task and throws a ChannelException to indicate that no message was persisted. However, it will be possible that a message was persisted. This case should only occur when the channel was halted and therefore the user has no guarantee of the state of the messages. In the worst case, the sender would resend the message and we would process it twice. 

Moved the destinationChainExecutor's catch block for RejectedExecutionExceptions to the correct line.

Flipped the boolean logic of processLock so it makes more logical sense.

Removed some debug code in VmSender
Issue: MIRTH-2250