MIRTH-2552 MIRTH-2499
Fixed MigratableConverter.unmarshal() so that it reads the object's version number from the top-level document element, rather than each migratable element since it is now only on the document element.

Also fixed MigratableConverter.unmarshal() so that it reads the current object's class type from the unmarshalling context, rather than trying to determine it from the element's name (this would not work on child elements since the child element names are field names and not class names).

Moved elementToXml() and elementFromXml() out of XmlUtil and into a new MigrationUtil class, since XmlUtil is exposed to the User through the TemplateValueReplacer, these methods didn't really need to be there.

Removed hardcoded "3.0.0" version number and made the Server and Client each initialize ObjectXMLSerializer with the current server version on startup.

Moved compareVersions() and normalizeVersion() from MigratableConverter into MigrationUtil.

Restored previous behavior when importing channels in the message browser so that it prompts the user when importing a channel from a previous version of Mirth Connect.