MIRTH-1700 MIRTH-2499
Updated the XStream lib to 1.4.4 on Client and Donkey.

Cleaned up error/exception handling with object serialization and migration. If one or more serialized objects in the database fails to deserialize or migrate to the current version, an error will be logged and the offending object will be ignored (channels, code templates and alerts).

Renamed the previous SerializerException to XmlSerializerException since it is referenced by the XmlSerializer interface. Added a new SerializerException that is intended to be used by implementations of the Serializer interface in Donkey.

ObjectXMLSerializer now throws a SerializerException when serialization/deserialization or migration fails, instead of returning null. This is now consistent with the fact that XStream itself throws a runtime exception when it encounters an error when serializing or deserializing.

Refactored and rearranged code in DefaultMigrationController, Migrate2_0_0 and Migrate3_0_0.